fixed shellsort

+++ 1/shellSort.py
@@ -3,6 +3,7 @@ def przerwa(tablica):
     while h<len(tablica):
         h=3*h+1
     return h
+
 def shellSort(tablica):
     h=przerwa(tablica)
     while (h):
@@ -13,5 +14,5 @@ def shellSort(tablica):
                 tablica[j+h]=tablica[j]
                 j-=h
             tablica[j+h]=tymczasowa
-        h=(h-1)/3
+        h=int((h-1)/3)
     return tablica
